feat(sdk): extract @clan-world/sdk package — unified codegen surface (closes #476)
Stage-1 plan validated at xhigh reasoning. Stage-2 implementation per Liam's 5 decisions on the open questions.

What changed:
- Add packages/sdk/ as new workspace package
- Move canonical Convex schema to packages/sdk/convex/schema.ts; apps/server/convex/schema.ts re-exports (no file-copy fallback needed — re-export survived server typecheck + Convex codegen)
- Move generated constants/enums canonical output to packages/sdk/src/generated/
- Replace packages/shared/src/generated/{constants,enums}.ts with compatibility re-exports from SDK (follow-up #488 will remove these once consumers migrate)
- Update packages/shared/src/types.ts to derive snapshot-facing types from @clan-world/sdk/convex
- Stub out packages/sdk/src/{solana,wormhole}/ subpaths for future cross-chain protocol integrations
- Wire SDK CI freshness checks into typechain-ci.yml

Subpaths:
- @clan-world/sdk — barrel for stable shared types
- @clan-world/sdk/contracts — re-exports @clan-world/contract-types
- @clan-world/sdk/constants + /enums — canonical generated chain mirrors
- @clan-world/sdk/convex — Doc<>, Id<>, DataModel
- @clan-world/sdk/convex/schema — Convex schema source
- @clan-world/sdk/solana — reserved (future Anchor IDLs)
- @clan-world/sdk/wormhole — reserved (future bridge types)

Deferred (filed as follow-ups):
- Narrow @clan-world/sdk/convex/api via makeFunctionReference
- Full web/mobile migrations (anyApi cleanup, WorldMap.tsx etc.)
- Native Kotlin codegen targeting SDK
- Absorbing @clan-world/contract-types
- Removing shared/generated/* compat re-exports (#488)

Verification (all green):
- pnpm --filter @clan-world/sdk typecheck
- pnpm --filter @clan-world/shared typecheck
- pnpm --filter @clan-world/server typecheck
- pnpm --filter @clan-world/contract-types typecheck
- pnpm --filter @clan-world/sdk codegen
- pnpm --filter @clan-world/server codegen
- pnpm test:chainclient-abi
- pnpm test:abi-parity
- node scripts/gen-constants.mjs --check && node scripts/gen-enums.mjs --check
