+++ apps/server/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@clan-world/contract-types": "workspace:*",
+    "@clan-world/sdk": "workspace:*",
     "@clan-world/shared": "workspace:*",
     "convex": "1.17.4",
     "viem": "2.48.4"
