+++ package.json
@@ -5,7 +5,7 @@
   "packageManager": "pnpm@10.33.2+sha512.a90faf6feeab71ad6c6e57f94e0fe1a12f5dcc22cd754db40ae9593eb6a3e0b6b12e3540218bb37ae083404b1f2ce6db2a4121e979829b4aff94b99f49da1cf8",
   "scripts": {
     "build": "turbo run build",
-    "codegen": "pnpm gen:constants && turbo run codegen",
+    "codegen": "pnpm gen:constants && pnpm gen:enums && turbo run codegen",
     "dev": "turbo run dev",
     "gen:chainclient-abi": "node scripts/gen-chainclient-abi.mjs",
     "gen:contract-abi": "node scripts/gen-contract-abi.mjs --build",
