+++ packages/sdk/convex.json
@@ -1,0 +1,3 @@
+{
+  "functions": "convex/"
+}
