+++ packages/sdk/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@clan-world/sdk",
+  "private": true,
+  "version": "2.2.0",
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./contracts": "./src/contracts.ts",
+    "./constants": "./src/generated/constants.ts",
+    "./enums": "./src/generated/enums.ts",
+    "./generated/*": "./src/generated/*.ts",
+    "./convex": "./src/convex/index.ts",
+    "./convex/schema": "./convex/schema.ts",
+    "./solana": "./src/solana/index.ts",
+    "./wormhole": "./src/wormhole/index.ts"
+  },
+  "scripts": {
+    "build": "tsc --noEmit",
+    "codegen": "convex codegen --typecheck disable",
+    "typecheck": "tsc --noEmit",
+    "lint": "echo 'lint stub'",
+    "clean": "rm -rf dist .turbo"
+  },
+  "dependencies": {
+    "@clan-world/contract-types": "workspace:*",
+    "convex": "1.17.4"
+  },
+  "devDependencies": {
+    "typescript": "5.9.3"
+  }
+}
