+++ packages/sdk/src/solana/index.ts
@@ -1,0 +1,1 @@
+// Reserved for Solana program IDLs (Anchor-generated) once consumed off-chain — see packages/solana-gold for the Rust source
