+++ packages/sdk/src/wormhole/index.ts
@@ -1,0 +1,1 @@
+// Reserved for Wormhole bridge type surfaces, cross-chain message schemas.
