+++ packages/sdk/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["src/**/*", "convex/**/*.ts", "convex/**/*.d.ts"]
+}
